chore: renaming files and fixing imports
